requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.10.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.4
sqlparse==0.4.3
tomli==2.0.1
